dont repeat build and upload if artifact is already present

+++ lib/tasks/build.rake
@@ -1,14 +1,21 @@
 desc 'Create a deployable artifact of the entire rails application'
 task :build do
+  rev_file = 'build/REVISION'
+  revision = `git rev-parse HEAD`
+  next if File.exists?(rev_file) && revision == File.read(rev_file)
+
   rm_rf 'build'
-  mkdir_p 'build'
+  mkdir 'build'
+
+  # compose contents
   sh 'git archive --format=tar HEAD | (cd build && tar -xf -)'
   Bundler.with_clean_env do
     sh 'cd build && RAILS_ENV=production bundle package --all-platforms'
   end
   sh 'rm -rf build/.bundle'
-  sh 'echo `git rev-parse HEAD` > build/REVISION'
-  sh 'rm -f build/raar.tar.gz'
+  File.write(rev_file, revision)
+
+  # create artifact
   sh 'cd build && ' \
      'tar -cvzf raar.tar.gz ' \
      '--exclude tmp --exclude test --exclude .travis.yml --exclude .rubocop.yml ' \
